Update public signing keys
The 2025 key is the current key to be used now, and the 2026 key is for emergencies if the staging server is compromised (which should never happen).

The keys don't have an expiration date set right now, but we can change them later to 8 and 16 years if we have a good reason.